[cms-v2] Add go back link for invoice/dcc details page
