+++ src/componentsv2/Router/index.js
@@ -1,2 +1,3 @@
 export { default as Router } from "./Router";
+export { default as GoBackLink } from "./GoBackLink";
 export * from "./RouterProvider";
